# blockchain: Add ancestor optimization to finding Ancestor