This mimics the behavior of Ancestor() in bitcoind