@davecgh 2021-11-17
Commenting here because I just say @Roasbeef's tweet linking to it.  FYI I implemented something similar in dcrd in what is imo a simpler way and should be a simple backport:

https://github.com/decred/dcrd/pull/2010