@kcalvinalvin 2023-11-17
Reviving this as I noticed that the startup of a node is slowed down quite a bit by the blockindex loading.

Here's a pprof of the startup blockindex loading without this PR.
![IMG_0424](https://github.com/btcsuite/btcd/assets/37185887/6bae9d71-4d92-4830-8a2a-e549833ff31b)

Here's one with the PR.
<img width="1366" alt="image" src="https://github.com/btcsuite/btcd/assets/37185887/24576d76-787d-4210-93eb-20813f29b8d8">

Added a benchmark as well. The benchstat results were:

```
    goos: darwin
    goarch: arm64
    pkg: github.com/utreexo/utreexod/blockchain
               │     old.txt      │               new.txt                │
               │      sec/op      │    sec/op     vs base                │
    Ancestor-8   120819.301µ ± 5%   7.013µ ± 19%  -99.99% (p=0.000 n=10)
    
               │  old.txt   │            new.txt             │
               │    B/op    │    B/op     vs base            │
    Ancestor-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
    ¹ all samples are equal
    
               │  old.txt   │            new.txt             │
               │ allocs/op  │ allocs/op   vs base            │
    Ancestor-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
    ¹ all samples are equal
```